fix deployment hibernate validation

+++ src/main/resources/db/migration/V1_1_0__tables_creation.sql
@@ -1,6 +1,6 @@
 CREATE TABLE attendees
 (
-    attendee_id  SERIAL PRIMARY KEY,
+    attendee_id  BIGINT(20) PRIMARY KEY,
     first_name   varchar(30) NOT NULL,
     last_name    varchar(30) NOT NULL,
     title        varchar(40) NULL,
@@ -27,7 +27,7 @@ CREATE TABLE pricing_categories
 
 CREATE TABLE ticket_prices
 (
-    ticket_price_id       SERIAL PRIMARY KEY,
+    ticket_price_id       BIGINT(20) PRIMARY KEY,
     ticket_type_code      varchar(1)    NOT NULL REFERENCES ticket_types (ticket_type_code),
     pricing_category_code varchar(1)    NOT NULL REFERENCES pricing_categories (pricing_category_code),
     base_price            numeric(8, 2) NOT NULL
@@ -35,7 +35,7 @@ CREATE TABLE ticket_prices
 
 CREATE TABLE discount_codes
 (
-    discount_code_id SERIAL PRIMARY KEY,
+    discount_code_id BIGINT(20) PRIMARY KEY,
     discount_code    varchar(20)   NOT NULL,
     discount_name    varchar(30)   NOT NULL,
     discount_type    varchar(1)    NOT NULL,
@@ -44,16 +44,16 @@ CREATE TABLE discount_codes
 
 CREATE TABLE attendee_tickets
 (
-    attendee_ticket_id SERIAL PRIMARY KEY,
-    attendee_id        integer       NOT NULL REFERENCES attendees (attendee_id),
-    ticket_price_id    integer       NOT NULL REFERENCES ticket_prices (ticket_price_id),
-    discount_code_id   integer       NULL REFERENCES discount_codes (discount_code_id),
+    attendee_ticket_id BIGINT(20) PRIMARY KEY,
+    attendee_id        BIGINT(20)       NOT NULL REFERENCES attendees (attendee_id),
+    ticket_price_id    BIGINT(20)       NOT NULL REFERENCES ticket_prices (ticket_price_id),
+    discount_code_id   BIGINT(20)       NULL REFERENCES discount_codes (discount_code_id),
     net_price          numeric(8, 2) NOT NULL
 );
 
 CREATE TABLE time_slots
 (
-    time_slot_id         SERIAL PRIMARY KEY,
+    time_slot_id         BIGINT(20) PRIMARY KEY,
     time_slot_date       date                   NOT NULL,
     start_time           time  					NOT NULL,
     end_time             time  					NOT NULL,
@@ -62,7 +62,7 @@ CREATE TABLE time_slots
 
 CREATE TABLE sessions
 (
-    session_id          SERIAL PRIMARY KEY,
+    session_id          BIGINT(20) PRIMARY KEY,
     session_name        varchar(80)   NOT NULL,
     session_description varchar(1024) NOT NULL,
     session_length      integer       NOT NULL
@@ -70,27 +70,27 @@ CREATE TABLE sessions
 
 CREATE TABLE session_schedule
 (
-    schedule_id  SERIAL PRIMARY KEY,
-    time_slot_id integer     NOT NULL REFERENCES time_slots (time_slot_id),
-    session_id   integer     NOT NULL REFERENCES sessions (session_id),
+    schedule_id  BIGINT(20) PRIMARY KEY,
+    time_slot_id BIGINT(20)     NOT NULL REFERENCES time_slots (time_slot_id),
+    session_id   BIGINT(20)     NOT NULL REFERENCES sessions (session_id),
     room         varchar(30) NOT NULL
 );
 
 CREATE TABLE tags
 (
-    tag_id      SERIAL PRIMARY KEY,
+    tag_id      BIGINT(20) PRIMARY KEY,
     description varchar(30) NOT NULL
 );
 
 CREATE TABLE session_tags
 (
-    session_id integer NOT NULL REFERENCES sessions (session_id),
-    tag_id     integer NOT NULL REFERENCES tags (tag_id)
+    session_id BIGINT(20) NOT NULL REFERENCES sessions (session_id),
+    tag_id     BIGINT(20) NOT NULL REFERENCES tags (tag_id)
 );
 
 CREATE TABLE speakers
 (
-    speaker_id    SERIAL PRIMARY KEY,
+    speaker_id    BIGINT(20) PRIMARY KEY,
     first_name    varchar(30)   NOT NULL,
     last_name     varchar(30)   NOT NULL,
     title         varchar(40)   NOT NULL,
@@ -101,13 +101,13 @@ CREATE TABLE speakers
 
 CREATE TABLE session_speakers
 (
-    session_id integer NOT NULL REFERENCES sessions (session_id),
-    speaker_id integer NOT NULL REFERENCES speakers (speaker_id)
+    session_id BIGINT(20) NOT NULL REFERENCES sessions (session_id),
+    speaker_id BIGINT(20) NOT NULL REFERENCES speakers (speaker_id)
 );
 
 CREATE TABLE workshops
 (
-    workshop_id   SERIAL PRIMARY KEY,
+    workshop_id   BIGINT(20) PRIMARY KEY,
     workshop_name varchar(60)   NOT NULL,
     description   varchar(1024) NOT NULL,
     requirements  varchar(1024) NOT NULL,
@@ -117,12 +117,12 @@ CREATE TABLE workshops
 
 CREATE TABLE workshop_speakers
 (
-    workshop_id integer NOT NULL REFERENCES workshops (workshop_id),
-    speaker_id  integer NOT NULL REFERENCES speakers (speaker_id)
+    workshop_id BIGINT(20) NOT NULL REFERENCES workshops (workshop_id),
+    speaker_id  BIGINT(20) NOT NULL REFERENCES speakers (speaker_id)
 );
 
 CREATE TABLE workshop_registrations
 (
-    workshop_id        integer NOT NULL REFERENCES workshops (workshop_id),
-    attendee_ticket_id integer NOT NULL REFERENCES attendee_tickets (attendee_ticket_id)
+    workshop_id        BIGINT(20) NOT NULL REFERENCES workshops (workshop_id),
+    attendee_ticket_id BIGINT(20) NOT NULL REFERENCES attendee_tickets (attendee_ticket_id)
 );
